requirements.txt
Django==1.4.3
django-debug-toolbar==0.9.4
django-tastypie==0.9.11
mimeparse==0.1.3
python-dateutil==1.5
wsgiref==0.1.2
